+++ CryptoBotCore/CosmosDB/Model/AccumulationSummary.cs
@@ -10,8 +10,8 @@ namespace CryptoBotCore.CosmosDB.Model
         [JsonProperty(PropertyName = "id")]
         public Guid Id { get; set; }
         public string CryptoName { get; set; }
-        public double AccumulatedCryptoAmount { get; set; }
-        public double InvestedFiatAmount { get; set; }
+        public decimal AccumulatedCryptoAmount { get; set; }
+        public decimal InvestedFiatAmount { get; set; }
         public int Buys { get; set; }
     }
 }
